(pd) move test files to __tests__/ dir
